move useMergeRefs and remove overflow from storybook decorator
