+++ .changeset/nice-tips-hide.md
@@ -1,0 +1,5 @@
+---
+'@lg-tools/storybook-decorators': patch
+---
+
+Removes `overflow: auto` from `rootStyles`.
