+++ .changeset/polite-kids-admire.md
@@ -1,0 +1,5 @@
+---
+'@leafygreen-ui/hooks': patch
+---
+
+Adds `useMergeRefs` into `useMergeRef` directory
